use force RMS as convergence criteria instead of fmax
* use force RMS as convergence criteria instead of fmax
* fix one bug
